# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos y Cart",
        "description": "Documentación de la API para gestionar productos y ventas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Obtener el carrito del usuario",
                "description": "Devuelve el carrito del usuario, crea uno si no existe",
                "operationId": "86f867f72e674aacd82942dcf38ba77e",
                "responses": {
                    "200": {
                        "description": "Carrito obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID del carrito",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID del usuario",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Fecha de creación del carrito",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Fecha de última actualización",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "cart_items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID del artículo del carrito",
                                                        "type": "integer"
                                                    },
                                                    "cart_id": {
                                                        "description": "ID del carrito asociado",
                                                        "type": "integer"
                                                    },
                                                    "product_id": {
                                                        "description": "ID del producto",
                                                        "type": "integer"
                                                    },
                                                    "quantity": {
                                                        "description": "Cantidad de producto en el carrito",
                                                        "type": "integer"
                                                    },
                                                    "product": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID del producto",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "Nombre del producto",
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "description": "Estado de activación del producto",
                                                                "type": "integer"
                                                            },
                                                            "price": {
                                                                "description": "Precio del producto",
                                                                "type": "string"
                                                            },
                                                            "stock": {
                                                                "description": "Cantidad disponible en inventario",
                                                                "type": "integer"
                                                            },
                                                            "ean": {
                                                                "description": "Código EAN del producto",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encuentra el carrito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Agregar producto al carrito",
                "description": "Agrega un producto al carrito del usuario",
                "operationId": "de31cc51185f893daa9bc71f17a59c60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID del producto",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Cantidad del producto",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha guardo correctamente el carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID del carrito",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID del usuario",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Fecha de creación del carrito",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Fecha de última actualización",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "cart_items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID del artículo del carrito",
                                                        "type": "integer"
                                                    },
                                                    "cart_id": {
                                                        "description": "ID del carrito asociado",
                                                        "type": "integer"
                                                    },
                                                    "product_id": {
                                                        "description": "ID del producto",
                                                        "type": "integer"
                                                    },
                                                    "quantity": {
                                                        "description": "Cantidad de producto en el carrito",
                                                        "type": "integer"
                                                    },
                                                    "product": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID del producto",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "Nombre del producto",
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "description": "Estado de activación del producto",
                                                                "type": "integer"
                                                            },
                                                            "price": {
                                                                "description": "Precio del producto",
                                                                "type": "string"
                                                            },
                                                            "stock": {
                                                                "description": "Cantidad disponible en inventario",
                                                                "type": "integer"
                                                            },
                                                            "ean": {
                                                                "description": "Código EAN del producto",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al agregar producto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/remove/{itemId}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Eliminar producto del carrito",
                "description": "Elimina un producto del carrito del usuario",
                "operationId": "07c2761625ef40eeba86661b19582ab3",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "ID del artículo a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/taxes": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Calcular impuestos del carrito",
                "description": "Calcula los impuestos basados en los productos del carrito",
                "operationId": "b4f5fffba5a53f2ce2ae9abddd716e1c",
                "responses": {
                    "200": {
                        "description": "Impuestos calculados correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "taxes": {
                                            "description": "El valor de los impuestos calculados",
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación para acceder a la API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}